refactor(exec): split CodexClient session workflow
